# api/openapi.yaml
openapi: 3.1.0
info:
  title: XLeaks API
  description: Decentralized P2P social platform API.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local node
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # Identity
...
    SwitchIdentityRequest:
      type: object
      properties:
        passphrase:
          type: string
      required: [passphrase]
    SwitchIdentityResponse:
      type: object
      properties:
        status:
          type: string
        pubkey:
          type: string
        address:
          type: string
    ExportIdentityResponse:
      type: object
      properties:
        pubkey:
          type: string
        address:
          type: string
        encrypted_key:
          type: string

    # Posts
...
    UpdateProfileRequest:
      type: object
      properties:
        display_name:
          type: string
          maxLength: 50
        bio:
          type: string
          maxLength: 500
        website:
          type: string
          maxLength: 200
        avatar_cid:
          type: string
        banner_cid:
          type: string
      required: [display_name]
    ProfileResponse:
      type: object
      properties:
        pubkey:
          type: string
        display_name:
          type: string
          maxLength: 50
        bio:
          type: string
          maxLength: 500
        website:
          type: string
          maxLength: 200
        avatar_cid:
          type: string
        banner_cid:
          type: string
        version:
          type: integer
        timestamp:
          type: integer

    # Follow
...
    SendDMRequest:
      type: object
      properties:
        content:
          type: string
      required: [content]
    DMResponse:
      type: object
      properties:
        id:
          type: string
        author:
          type: string
        recipient:
          type: string
        timestamp:
          type: integer
...
    MessageEntry:
      type: object
      properties:
        id:
          type: string
        author:
          type: string
        recipient:
          type: string
        content:
          type: string
        timestamp:
          type: integer
        read:
          type: boolean

    # Notifications
...
  parameters:
...
    PubkeyParam:
      name: pubkey
      in: path
      required: true
      schema:
        type: string
      description: Hex-encoded public key
    BeforeParam:
      name: before
      in: query
      schema:
        type: integer
      description: Cursor timestamp for pagination (milliseconds)
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
      description: Number of items to return
paths:
...
  /api/identity/list:
    get:
      tags: [Identity]
      summary: List all stored identities
      operationId: listIdentities
      responses:
        '200':
          description: List of identities
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/identity/switch/{pubkey}:
    put:
      tags: [Identity]
      summary: Switch to a different identity
      operationId: switchIdentity
      parameters:
        - $ref: '#/components/parameters/PubkeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwitchIdentityRequest'
      responses:
        '200':
          description: Identity switched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwitchIdentityResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Wrong passphrase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/identity/export:
    post:
      tags: [Identity]
      summary: Export the active identity
      operationId: exportIdentity
      responses:
        '200':
          description: Exported identity file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportIdentityResponse'
        '404':
          description: No active identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Posts
...
  /api/profile:
    get:
      tags: [Profile]
      summary: Get own profile
      operationId: getOwnProfile
      responses:
        '200':
          description: Profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Identity not unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Profile]
      summary: Update own profile
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Identity not unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/dm/{pubkey}:
    get:
      tags: [DMs]
      summary: Get conversation with a user
      operationId: getConversation
      parameters:
        - $ref: '#/components/parameters/PubkeyParam'
        - $ref: '#/components/parameters/BeforeParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageEntry'
        '400':
          description: Invalid pubkey
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Identity not unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [DMs]
      summary: Send a direct message
      operationId: sendDM
      parameters:
        - $ref: '#/components/parameters/PubkeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendDMRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DMResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Identity not unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Notifications
...